Move context stuff to its own file
